:fire: Do not get a token during call to recognize with websocket
 * Remove the need to call to the authorization service when creating a
WebSocket.
 * Increase the WebSocket frame size to 16 kb
 * Reuse the OkHttpClient for WebSocket calls
